pyproject.toml
[tool.poetry]
name = "fake-nft-pfp"
version = "0.1.0"
description = ""
authors = ["sindarknave <ngwingkui@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.0.0"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
